Fix rubocop config to make heredocs possible
Apparently the puppet community prefers to enforce trailing comments on
multiline functions in ruby, which makes it impossible to do this:

    it do
      is_expected.to contain_file('/etc/fwknop/fwknopd.conf').with_content(
        <<~FWKNOPD
          # Managed by puppet.
          PCAP_INTF #{facts[:networking]['primary']}
        FWKNOPD
      )
    end

Rubocop failed because it wanted me to add a trailing comma, but

           # Managed by puppet.
           PCAP_INTF #{facts[:networking]['primary']}
    -    FWKNOPD
    +    FWKNOPD,
       )
     end

breaks the heredoc and

           # Managed by puppet.
           PCAP_INTF #{facts[:networking]['primary']}
         FWKNOPD
    +    ,
       )
     end

is invalid ruby.

So I'm changing our setting for trailing commas to the setting that
standard ruby uses.